cleaner code. Added loading state to catch boundary
