+++ src/css/global-styles.css
@@ -39,7 +39,6 @@
 }
 
 .footer {
-  position: absolute;
   bottom: 0;
   background-color: black;
   text-align: center;
